% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsesqlr.R
\name{.agregar_boletins}
\alias{.agregar_boletins}
\title{Agrega o boletim de urna WEB por municipio/cargo/votavel}
\usage{
.agregar_boletins(ano, uf)
}
\description{
Agrega o boletim de urna WEB por municipio/cargo/votavel
}
\keyword{internal}
