% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsesqlr.R
\name{.carregar_um}
\alias{.carregar_um}
\title{Carga de um escopo (tipo, ano, uf) — nucleo do \code{\link[=tsesqlr_carregar]{tsesqlr_carregar()}}}
\usage{
.carregar_um(tipo, ano, uf, con, refrescar)
}
\description{
Carga de um escopo (tipo, ano, uf) — nucleo do \code{\link[=tsesqlr_carregar]{tsesqlr_carregar()}}
}
\keyword{internal}
