% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsesqlr.R
\docType{data}
\name{.padrao_colunas}
\alias{.padrao_colunas}
\title{Colunas canonicas por tabela (esquema estavel entre anos/ciclos)}
\format{
An object of class \code{list} of length 4.
}
\usage{
.padrao_colunas
}
\description{
Colunas canonicas por tabela (esquema estavel entre anos/ciclos)
}
\keyword{internal}
