% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsesqlr.R
\docType{data}
\name{.tipos_colunas}
\alias{.tipos_colunas}
\title{Tipos canonicos das colunas (evita NA logico virar BOOLEAN no banco)}
\format{
An object of class \code{character} of length 29.
}
\usage{
.tipos_colunas
}
\description{
Tipos canonicos das colunas (evita NA logico virar BOOLEAN no banco)
}
\keyword{internal}
