% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsesqlr.R
\docType{data}
\name{.ufs_br}
\alias{.ufs_br}
\title{Siglas das 27 UFs (sem ZZ)}
\format{
An object of class \code{character} of length 27.
}
\usage{
.ufs_br
}
\description{
Siglas das 27 UFs (sem ZZ)
}
\keyword{internal}
