% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsesqlr.R
\name{.padronizar}
\alias{.padronizar}
\title{Completa/ordena colunas do data.frame no esquema canonico}
\usage{
.padronizar(dados, colunas, ano)
}
\description{
Completa/ordena colunas do data.frame no esquema canonico
}
\keyword{internal}
